lab13/openapiplacerecordv4/requirements.txt
flask-restful
flask-restful-swagger-3
flask-cors